plant-disease-prediction-cnn-deep-leanring-project-main/app/requirements.txt

pillow
numpy==1.26.3
streamlit==1.30.0
tensorflow==2.15.0